Don't fail noperm tests if ran as root.
Cirrus does so on some platforms, and this test is already done elsewhere
thus I don't bother to muck with adduser & co.

+++ tests/d-file-noperm.t
@@ -1,3 +1,4 @@
+[ `whoami` != root ] || exit 42
 cp -p $F file
 $TOOL file
 rm -f file
